requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
commonmark==0.9.1
decorator==4.4.2
flake8==3.8.3
greenlet==1.1.2
gTTS==2.2.4
idna==3.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
mccabe==0.6.1
moviepy==1.0.3
mutagen==1.45.1
mypy-extensions==0.4.3
numpy==1.22.3
pathspec==0.8.0
Pillow==9.1.1
playwright==1.22.0
praw==7.6.0
prawcore==2.3.0
proglog==0.1.10
pycodestyle==2.6.0
pyee==8.1.0
pyflakes==2.2.0
Pygments==2.12.0
python-dotenv==0.20.0
pytube==12.1.0
regex==2020.10.15
requests==2.27.1
rich==12.4.4
six==1.16.0
toml==0.10.1
tqdm==4.64.0
# typed-ast==1.4.1
typing_extensions==4.2.0
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.2
websockets==10.1
